+++ Demo/DemoPresenter.swift
@@ -1,39 +1,34 @@
 import Foundation
 import CleanArchitecture
 
-class DemoPresenter:Presenter {
-    var viewModels:ViewModels!
-    var interactor:DemoInteractor!
-    
-    required init() { }
-    
+class DemoPresenter:Presenter<DemoInteractor> {
     func updatedRate(string:String) {
         let exchange = interactor.updateRateWith(string:string)
         var property = UsdViewModel()
         property.amount = exchange.usd
-        viewModels.update(viewModel:property)
+        update(viewModel:property)
     }
     
     func updatedEuro(string:String) {
         let exchange = interactor.updateEuroWith(string:string)
         var property = UsdViewModel()
         property.amount = exchange.usd
-        viewModels.update(viewModel:property)
+        update(viewModel:property)
     }
     
     func updatedUsd(string:String) {
         let exchange = interactor.updateUsdWith(string:string)
         var property = EuroViewModel()
         property.amount = exchange.euro
-        viewModels.update(viewModel:property)
+        update(viewModel:property)
     }
     
-    func didAppear() {
+    override func didAppear() {
         let exchange = interactor.initialExchanger()
         var property = DemoViewModel()
         property.rate = exchange.rate
         property.euro = exchange.euro
         property.usd = exchange.usd
-        viewModels.update(viewModel:property)
+        update(viewModel:property)
     }
 }