+++ Source/Presenter.swift
@@ -1,23 +1,40 @@
 import Foundation
 
-public protocol Presenter:InteractorDelegate {
-    associatedtype I:Interactor
-    var interactor:I! { get set }
-    var viewModels:ViewModels! { get set }
+open class Presenter<I:Interactor>:Delegate {
+    open var interactor:I!
+    private(set) var viewModels:[ObjectIdentifier:Any] = [:]
     
-    init()
-    func didLoad()
-    func willAppear()
-    func didAppear()
+    public required init() { }
+    open func didLoad() { }
+    open func willAppear() { }
+    open func didAppear() { }
+    open func shouldUpdate() { }
+    
+    public func update<V>(viewModel:V!) {
+        var item = Item<V>()
+        item.viewModel = viewModel
+        item.observer = cached().observer
+        viewModels[ObjectIdentifier(V.self)] = item
+        DispatchQueue.main.async { item.observer?(viewModel) }
+    }
+    
+    public func viewModel<V>(observer:@escaping((V) -> Void)) {
+        var item:Item<V> = cached()
+        item.observer = observer
+        viewModels[ObjectIdentifier(V.self)] = item
+    }
+    
+    private func cached<V>() -> Item<V> {
+        guard let item = viewModels[ObjectIdentifier(V.self)] as? Item<V> else { return Item<V>() }
+        return item
+    }
 }
 
-public extension Presenter {
-    func didLoad() { }
-    func willAppear() { }
-    func didAppear() { }
-    func shouldUpdate() { }
-}
-
-public protocol InteractorDelegate:AnyObject {
+public protocol Delegate:AnyObject {
     func shouldUpdate()
+}
+
+private struct Item<V> {
+    var observer:((V) -> Void)?
+    var viewModel:V!
 }