+++ Source/View.swift
@@ -1,20 +1,18 @@
 import UIKit
 
-open class View<P:Presenter>:UIViewController {
+open class View<I, P>:UIViewController where P:Presenter<I> {
     open var presenter:P
     
     public required init() {
         presenter = P()
-        presenter.interactor = P.I()
+        presenter.interactor = I()
         presenter.interactor.delegate = presenter
         super.init(nibName:nil, bundle:nil)
-        presenter.viewModels = ViewModels()
     }
     
     public init(presenter:P) {
         self.presenter = presenter
         super.init(nibName:nil, bundle:nil)
-        presenter.viewModels = ViewModels()
     }
     
     public required init?(coder:NSCoder) { return nil }