+++ Tests/MockViewModel.swift
@@ -1,10 +1,5 @@
 import Foundation
-import CleanArchitecture
 
-struct MockViewModel:ViewModel {
-    var name:String
-    
-    init() {
-        name = String()
-    }
+struct MockViewModel {
+    var name = String()
 }